requirements_test.txt
# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component>=0.3.0
aioresponses>=0.7.0
ruff>=0.1.0
mypy>=1.0.0
black>=23.0.0
codecov>=2.0.0